prepare Lumi for extraction
