+++ examples/led-matrix/lib/Matrix/Matrix.Character.h
@@ -1,15 +1,13 @@
-#ifndef LUMI_COUNT_H
-#define LUMI_COUNT_H
+#ifndef MATRIX_CHARACTER_H
+#define MATRIX_CHARACTER_H
 
 #include <WEMOS_Matrix_GFX.h>
 #include <Lumi.Pattern.h>
 
-#define GRID_SIZE 8
-
-namespace Lumi {
-  class Count : public Pattern {
+namespace Matrix {
+  class Character : public Lumi::Pattern {
     public:
-      Count(MLED &matrix);
+      Character(MLED &matrix);
       void update(const uint8_t progress) override;
     private:
       MLED &matrix;
