+++ examples/led-matrix/lib/Matrix/Matrix.Count.h
@@ -1,13 +1,15 @@
-#ifndef LUMI_CHARACTER_H
-#define LUMI_CHARACTER_H
+#ifndef MATRIX_COUNT_H
+#define MATRIX_COUNT_H
 
 #include <WEMOS_Matrix_GFX.h>
 #include <Lumi.Pattern.h>
 
-namespace Lumi {
-  class Character : public Pattern {
+#define GRID_SIZE 8
+
+namespace Matrix {
+  class Count : public Lumi::Pattern {
     public:
-      Character(MLED &matrix);
+      Count(MLED &matrix);
       void update(const uint8_t progress) override;
     private:
       MLED &matrix;
