+++ examples/led-rgb/lib/Led/Led.LEDPattern.h
@@ -1,11 +1,11 @@
-#ifndef LUMI_LED_PATTERN_H
-#define LUMI_LED_PATTERN_H
+#ifndef LED_LED_PATTERN_H
+#define LED_LED_PATTERN_H
 
 #include <Lumi.Pattern.h>
 #include <FastLED.h>
 
-namespace Lumi {
-  class LEDPattern : public Pattern {
+namespace Led {
+  class LEDPattern : public Lumi::Pattern {
     public:
       LEDPattern(CRGB *leds, const uint8_t numLeds);
     protected:
