+++ examples/led-rgb/lib/Led/Led.Off.cpp
@@ -1,6 +1,6 @@
-#include <Lumi.Off.h>
+#include <Led.Off.h>
 
-namespace Lumi {
+namespace Led {
   void Off::update(const uint8_t progress) {
     for(int i = 0; i < numLeds; i++ ) {
       leds[i] = CRGB(0, 0, 0);
