+++ examples/led-rgb/src/main.cpp
@@ -6,10 +6,10 @@ using namespace Bricks;
 // Local
 #include <FastLED.h>
 #include <Lumi.Sequencer.h>
-#include <Lumi.Off.h>
-#include <Lumi.Glow.h>
-#include <Lumi.Spin.h>
-#include <Lumi.Wave.h>
+#include <Led.Off.h>
+#include <Led.Glow.h>
+#include <Led.Spin.h>
+#include <Led.Wave.h>
 
 #define NUM_LEDS 7
 #define LEDS_PIN D2
@@ -35,10 +35,10 @@ void setup() {
   FastLED.addLeds<WS2811, LEDS_PIN, GRB>(leds, NUM_LEDS).setCorrection(TypicalLEDStrip);
 
   // Configure animations
-  sequencer.patterns[0] = new Lumi::Off(leds, NUM_LEDS);
-  sequencer.patterns[1] = new Lumi::Glow(leds, NUM_LEDS);
-  sequencer.patterns[2] = new Lumi::Spin(leds, NUM_LEDS);
-  sequencer.patterns[3] = new Lumi::Wave(leds, NUM_LEDS);
+  sequencer.patterns[0] = new Led::Off(leds, NUM_LEDS);
+  sequencer.patterns[1] = new Led::Glow(leds, NUM_LEDS);
+  sequencer.patterns[2] = new Led::Spin(leds, NUM_LEDS);
+  sequencer.patterns[3] = new Led::Wave(leds, NUM_LEDS);
 }
 
 void loop() {
